Bug Fix
It was only comparing the first element of the mouse position. This was affecting both the purchase buttons and the clicking of the balloons.  The bug where it clicked multiple at once should be fixed now, although there's still an issue where it register the mouse position instantly, so if you move the mouse away after clicking before the tick goes, it might register your mouse as somewhere else.